refactor: 구조 domain형 구조로 전체 리팩토링 진행, main과 test 구조 통일
