+++ src/main/java/domain/car/Car.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.car;
 
 public class Car {
