+++ src/main/java/domain/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.generator;
 
 public interface NumberGenerator    {
     int getNumber();
