+++ src/test/java/domain/car/CarTest.java
@@ -1,4 +1,5 @@
-import domain.Car;
+package domain.car;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
