+++ src/test/java/domain/generator/RandomNumberGeneratorTest.java
@@ -1,4 +1,5 @@
-import domain.RandomNumberGenerator;
+package domain.generator;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
